Restructured pages into modules
