+++ react/src/pages/Auth/LoginPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import LoginForm from "../components/LoginForm";
-import '../style/auth.css';
+import LoginForm from "../../components/LoginForm";
+import './auth.css';
 import {Link} from "react-router-dom";
 
 const LoginPage = () => (
