+++ react/src/pages/HomePage/index.js
@@ -1,4 +1,4 @@
-import Footer from "../components/Footer";
+import Footer from "../../components/Footer";
 import * as React from 'react';
 import {styled, createTheme, ThemeProvider} from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
@@ -13,7 +13,7 @@ import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
 import PeopleIcon from '@mui/icons-material/People';
 import BarChartIcon from '@mui/icons-material/BarChart';
 import LayersIcon from '@mui/icons-material/Layers';
-import Sidebar from "../components/SideBar";
+import Sidebar from "../../components/SideBar";
 import {Link, Outlet} from "react-router-dom";
 
 function ListItemTextWhite({primary}) {
