+++ react/src/pages/NewSurveyPage/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
-import NewSurvey from "../components/NewSurvey";
+import NewSurvey from "../../components/NewSurvey";
 
-//code transferred to component**
 
 const NewSurveyPage = () => (
     <>
