+++ react/src/pages/PrivacyPage/index.js
